fix: correct type import by moving new type to interfaces folder
