+++ server/entity/MediaRequest.ts
@@ -20,11 +20,11 @@ import TheMovieDb from '../api/themoviedb';
 import { ANIME_KEYWORD_ID } from '../api/themoviedb/constants';
 import { MediaRequestStatus, MediaStatus, MediaType } from '../constants/media';
 import { getRepository } from '../datasource';
+import type { MediaRequestBody } from '../interfaces/api/requestInterfaces';
 import notificationManager, { Notification } from '../lib/notifications';
 import { Permission } from '../lib/permissions';
 import { getSettings } from '../lib/settings';
 import logger from '../logger';
-import type { MediaRequestBody } from '../routes/request';
 import Media from './Media';
 import SeasonRequest from './SeasonRequest';
 import { User } from './User';
